refactor: divide DataList index.tsx to be more lightweight
